Add files via upload
This is  a project on employee attrition in a company that makes prediction on the probability of employees leaving or staying in the company. An attrition of 0 indicates the employee stays while an attrition of 1 indicates that the employee leaves the company. This project was accomplished using Naive Bayesian machine learning algorithm to make the predictions by training the model on training dataset and testing with testing dataset.