Add files via upload
This is a project on the study of churn cases in a Telecom company using Logistic Regression model. We try to predict whether the customer leaves the Telecom company or stays based on several factors or features of the dataset. A churn of 0 implies the customer stays and a churn of 1 implies the customer leaves